refactor: harden runtime and editor ownership boundaries
- Split runtime rendering, asset, scene, UI, dynamic API, and plugin extension code into clearer owner modules.
- Replace legacy render pipeline helpers with focused graph, material, mesh, shadow, post-process, visibility, and shader source structures.
- Rework editor workbench layout assets, host projections, command routing, template painting, and retained host callback coverage.
- Update plugin export, standalone build, asset importer, and audit tooling documentation and support code.
- Expand structure, budget, lock-poison, render product, UI, plugin manifest, and editor host tests around the new module layout.
- Refresh runtime, render, editor layout, plugin, architecture, and code review plan docs to match the reorganized workspace.

+++ zircon_editor/src/ui/retained_host/host_contract/native_pointer/button_dispatch/chrome_press/tabs/host_page.rs
@@ -3,12 +3,12 @@ use crate::ui::retained_host::host_contract::window::UiHostWindow;
 
 pub(in crate::ui::retained_host::host_contract::native_pointer::button_dispatch::chrome_press) fn dispatch_host_page_tab_press(
     ui: &UiHostWindow,
-    index: usize,
+    index: i32,
     tab_x: f32,
     tab_width: f32,
     local_x: f32,
     local_y: f32,
 ) {
     ui.global::<UiHostContext>()
-        .invoke_host_page_pointer_clicked(index as i32, tab_x, tab_width, local_x, local_y);
+        .invoke_host_page_pointer_clicked(index, tab_x, tab_width, local_x, local_y);
 }
